ios: fix Valve/ObjC BOOL collision in the new .mm files
CI caught this on the all-features build:
  objc.h:78 error: typedef redefinition with different types
                   ('bool' vs 'int')
  basetypes.h:146 note: previous definition is here -> typedef int BOOL

Valve defines BOOL as int, Objective-C defines it as bool, and any
translation unit that sees both fails. The repo solved this once before in
commit 14f6ad3d by keeping UIKit out of glmrendererinfo and bridging with
extern "C"; this applies the same structure to the new code.

  * snd_dev_ios_phase.mm split into snd_dev_ios_phase.cpp (Valve headers,
    no ObjC) plus snd_dev_ios_phase_backend.mm (AVFoundation, no Valve
    headers), talking over a plain C bridge header.
  * haptics_ios.mm and ios_metal_device.mm no longer include tier0
    headers; diagnostics go through printf.
  * ios_metal_device.h uses unsigned long long rather than Valve's uint64,
    since it is included from the ObjC side.
  * MTLGPUFamilyApple8/9 dropped - not declared by the Xcode 15.4 SDK.
  * Added scripts/check_objc_bool.py so this class of error is caught in
    under a second instead of a 10 minute CI round trip.

Baseline (all flags off) built green on this base, confirming the opt-in
changes are inert when disabled.

+++ appframework/ios_metal_device.h
@@ -50,7 +50,9 @@ struct IOSMetalDeviceInfo_t
 {
 	char	m_szDeviceName[ 128 ];
 
-	// Highest Apple GPU family supported (1..9); 0 if unknown.
+	// Highest Apple GPU family supported (1..7); 0 if unknown.
+	// Capped at 7 because Xcode 15.4's SDK is the build floor here and
+	// MTLGPUFamilyApple8/9 are not guaranteed to be declared.
 	int		m_nAppleGPUFamily;
 
 	// Largest 2D texture dimension the hardware accepts.
@@ -59,7 +61,9 @@ struct IOSMetalDeviceInfo_t
 	// Recommended working set in bytes. Exceeding it makes iOS start
 	// evicting our resources, which is the usual cause of texture
 	// popping on memory-constrained devices.
-	uint64	m_nRecommendedMaxWorkingSetSize;
+	// Plain C type: this header is included from an ObjC translation unit
+	// that must not pull in Valve's basetypes.h.
+	unsigned long long	m_nRecommendedMaxWorkingSetSize;
 
 	bool	m_bSupportsFamilyApple4;	// A11 and later
 	bool	m_bSupportsFamilyApple7;	// A14/M1 and later
